Moved keyword tokenizer factory back to server, because normalizers directly depend on it.This should be addressed on a follow up change.
